requirements.txt
python-dotenv
agno
openai
pandas
matplotlib
seaborn
openpyxl
xlrd
xlwings
numpy
scipy
plotly
sentence-transformers
faiss-cpu
pillow
requests
beautifulsoup4
lxml